APP-4531 Move interface in a seperate file and use 'defaultProps'
