+++ src/components/time-picker/interfaces.ts
@@ -1,3 +1,6 @@
+import { HasTooltipProps } from '../tooltip/interfaces';
+import { HasValidationProps } from '../validation/interfaces';
+
 export type TimePickerValue = string;
 
 interface DisabledExactTime {
@@ -6,7 +9,24 @@ interface DisabledExactTime {
 
 interface DisabledTimeRange {
   from: string;
-  to : string;
+  to: string;
 }
 
 export type DisabledTime = DisabledExactTime | DisabledTimeRange;
+
+export type TimePickerProps = {
+  id?: string;
+  disabled?: boolean;
+  disabledTimes?: DisabledTime | Array<DisabledTime>;
+  format?: string;
+  label?: string;
+  min?: string;
+  max?: string;
+  name?: string;
+  placeholder?: string;
+  step?: number;
+  strict?: boolean;
+  value?: string;
+  showRequired?: boolean;
+} & HasValidationProps<TimePickerValue> &
+  HasTooltipProps;
